Section3-CloudBuild-CICD/cloudrun-ml-models/coupon-recommendations-v2/requirements.txt
Flask==2.1.0
gunicorn==20.1.0
pandas
pyarrow
xgboost==1.6.2
category_encoders
google-cloud-storage
